requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
flask-sqlalchemy>=3.0.0
requests>=2.31.0
numpy>=1.21.0
pandas>=1.5.0
scipy>=1.9.0
openai>=1.0.0
python-dotenv>=0.19.0
marshmallow>=3.19.0